tooling: extract ci-watch engine with shared options/result contract
